+++ smokesim/particle.py
@@ -1,26 +1,24 @@
+from smokesim.base import BaseSim, BaseProperty
+from smokesim.constants import CLOUD_MASK
+
 import pygame
 import cv2
-import random
 import numpy as np
-from dataclasses import dataclass
-from smokesim.constants import CLOUD_MASK
 from typing import Optional, Tuple
 
 
-
-@dataclass
-class Sprite:
+class Sprite(BaseProperty):
     """
     A dataclass to represent a sprite.
     """
+
     width: int = 20
     height: int = 20
     opacity_mask: np.ndarray = CLOUD_MASK
     color: tuple = (255, 255, 255)
 
 
-@dataclass
-class ParticleProperty:
+class ParticleProperty(BaseProperty):
     """
     A dataclass to represent the arguments for a particle.
 
@@ -43,15 +41,16 @@ class ParticleProperty:
     - fade_speed (int, optional): The speed at which the particle fades. Defaults to 1.
     - alpha (int, optional): The alpha of the particle. Defaults to 255.
     """
+
     vx: Optional[float] = None
     startvy: Optional[float] = None
     scale: Optional[float] = None
     lifetime: Optional[int] = None
     age: int = 0
-    min_vx: float = -4/100
-    max_vx: float = 4/100
-    min_vy: float = -4/10
-    max_vy: float = -1/10
+    min_vx: float = -4 / 100
+    max_vx: float = 4 / 100
+    min_vy: float = -4 / 10
+    max_vy: float = -1 / 10
     min_scale: int = 20
     max_scale: int = 40
     min_lifetime: float = 2000
@@ -60,12 +59,11 @@ class ParticleProperty:
     smoke_sprite_size: int = 20
     fade_speed: int = 1
     alpha: int = 255
-    seed:int = 100
+    # seed:int = 100
 
 
-class Particle:
-    def __init__(self, x: int, y: int, property: ParticleProperty
-                 ):
+class Particle(BaseSim):
+    def __init__(self, x: int, y: int, property: ParticleProperty):
         """
         A class to represent a particle. Particles are used to create smoke.
 
@@ -91,25 +89,37 @@ class Particle:
             - fade_speed (int, optional): The speed at which the particle fades. Defaults to 1.
             - alpha (int, optional): The alpha of the particle. Defaults to 255.
         """
-        self.seed = property.seed
-        random.seed(self.seed)
-        self.float_in_range = lambda start, end: start + np.random.random() * (end - start)
+        super().__init__(property)
+
+        # self.seed = property.seed
+        # self.float_in_range = lambda start, end, random_state: start + np.random.random() * (end - start)
         self.property = property
         self.x = x
         self.y = y
-        self.vx = property.vx if property.vx is not None else self.float_in_range(
-            property.min_vx, property.max_vx)
-        self.startvy = property.startvy if property.startvy is not None else self.float_in_range(
-            property.min_vy, property.max_vy)
-        self.scale = property.smoke_sprite_size if property.scale is not None else self.float_in_range(
-            property.min_scale, property.max_scale)
-        self.lifetime = property.lifetime if property.lifetime is not None else self.float_in_range(
-            property.min_lifetime, property.max_lifetime)
+        self.vx = (
+            property.vx
+            if property.vx is not None
+            else self.float_in_range(property.min_vx, property.max_vx)
+        )
+        self.startvy = (
+            property.startvy
+            if property.startvy is not None
+            else self.float_in_range(property.min_vy, property.max_vy)
+        )
+        self.scale = (
+            property.smoke_sprite_size
+            if property.scale is not None
+            else self.float_in_range(property.min_scale, property.max_scale)
+        )
+        self.lifetime = (
+            property.lifetime
+            if property.lifetime is not None
+            else self.float_in_range(property.min_lifetime, property.max_lifetime)
+        )
         self.age = property.age
         self.color = property.color
         self.smoke_sprite_size = property.smoke_sprite_size
-        self.final_scale = self.float_in_range(self.scale * 0.1,
-                                          self.scale*1.5)
+        self.final_scale = self.float_in_range(self.scale * 0.1, self.scale * 1.5)
         self.scale_step = (self.final_scale - self.scale) / self.lifetime
         self.vy = self.startvy
         self.alpha = property.alpha
@@ -127,19 +137,21 @@ class Particle:
         Returns:
         - pygame.Surface: The painted sprite.
         """
-        surface = pygame.Surface(
-            (sprite.width, sprite.height), pygame.SRCALPHA)
+        surface = pygame.Surface((sprite.width, sprite.height), pygame.SRCALPHA)
         pixels = pygame.PixelArray(surface)
         opacities = cv2.resize(
-            sprite.opacity_mask, (sprite.width, sprite.height), interpolation=cv2.INTER_NEAREST)
+            sprite.opacity_mask,
+            (sprite.width, sprite.height),
+            interpolation=cv2.INTER_NEAREST,
+        )
         for x in range(sprite.width):
             for y in range(sprite.height):
                 pixels[x, y] = (*sprite.color, opacities[x, y])
         del pixels
         surface = pygame.transform.smoothscale(
-            surface, (sprite.width // 2, sprite.height // 2))
-        surface = pygame.transform.smoothscale(
-            surface, (sprite.width, sprite.height))
+            surface, (sprite.width // 2, sprite.height // 2)
+        )
+        surface = pygame.transform.smoothscale(surface, (sprite.width, sprite.height))
         self.surface = surface
         return surface
 
@@ -151,7 +163,10 @@ class Particle:
         - pygame.Surface: The sprite.
         """
         self.sprite = Sprite(
-            color=self.color, width=self.smoke_sprite_size, height=self.smoke_sprite_size)
+            color=self.color,
+            width=self.smoke_sprite_size,
+            height=self.smoke_sprite_size,
+        )
         self.sprite_paint = self.paint_sprite(self.sprite)
         return self.sprite_paint
 
@@ -179,11 +194,12 @@ class Particle:
                 self.is_alive = False
 
             surface = pygame.transform.smoothscale(
-                self.surface, (self.sprite.width // 2, self.sprite.height // 2))
+                self.surface, (self.sprite.width // 2, self.sprite.height // 2)
+            )
             surface = pygame.transform.smoothscale(
-                self.surface, (self.sprite.width, self.sprite.height))
+                self.surface, (self.sprite.width, self.sprite.height)
+            )
             self.sprite_paint = surface
-            
 
     def draw(self, screen):
         """