+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## June 28, 2024
+### Added
+* Reproducibility based on random states.
+* Test case for random state. (Different state should lead to different result.)
+
+### Changed
+* 
+
 ## April 04, 2024
 ### Added
 * Properties for `Smoke` and `Particle`.